+++ changelog.md
@@ -3,7 +3,10 @@ Changelog
 
 # 1.0.1-SNAPSHOT
 * Added more methods to `OperatingSystemUtils` and `CommonUtils`
-* Started work on `EnvironmentUtils`
+* Added `FileUtils`
+* Improved `SerializeUtils` methods. More overloading methods.
+* Renamed `NothingAvailableException` to `NothingFoundException`
+* Get stacktrace method in `ReflectionUtils`. Small improvements.
 
 # 1.0.0
 First version